refactor(ui/web): migrate DashboardHeader components from @starboard
